+++ BlockManager.h
@@ -9,14 +9,14 @@ public:
 
 	void LoadBlocks();
 
-	Block& GetBlockById(uint32_t id);
+	Block& GetBlockById(uint16_t id);
 	Block& GetBlockByName(std::string name);
 
 private:
 	void BuildBlocksMap(std::vector<Block>& blocks);
 
 private:
-	std::unordered_map<uint32_t, Block> m_blocksById;
+	std::unordered_map<uint16_t, Block> m_blocksById;
 	std::unordered_map<std::string, Block> m_blocksByName;
 };
