+++ ChunkBlock.cpp
@@ -1,17 +1,17 @@
 #include "pch.h"
 #include "ChunkBlock.h"
 
-ChunkBlock::ChunkBlock(uint32_t id) noexcept :
+ChunkBlock::ChunkBlock(uint16_t id) noexcept :
 	m_id(id), m_dir(ChunkBlock::BlockDirection::North)
 {
 }
 
-ChunkBlock::ChunkBlock(uint32_t id, BlockDirection dir) noexcept :
+ChunkBlock::ChunkBlock(uint16_t id, BlockDirection dir) noexcept :
 	m_id(id), m_dir(dir)
 {
 }
 
-uint32_t ChunkBlock::GetId() const noexcept
+uint16_t ChunkBlock::GetId() const noexcept
 {
 	return m_id;
 }
